fixes and testing AI branch
