+++ pubspec.yaml
@@ -16,5 +16,4 @@ dev_dependencies:
     sdk: flutter
 
 flutter:
-
-  
+  uses-material-design: true
